Relocate Slippage functions
